# Port stale simple_keys fix to v2