Port @niemeyer's https://github.com/go-yaml/yaml/commit/e228e37189d388c1bff077892b97197221e505cb patch back to v2 to fix the issues highlighted in https://github.com/go-yaml/yaml/pull/537.

This PR includes the tests from https://github.com/go-yaml/yaml/pull/537 (as valid YAML this time!)